test: move AppLayoutToolbar SSR test to app-layout-toolbar directory
