+++ src/app-layout/__tests__/app-layout.ssr.test.tsx
@@ -9,7 +9,6 @@ import { renderToStaticMarkup } from 'react-dom/server';
 import { clearVisualRefreshState } from '@cloudscape-design/component-toolkit/internal/testing';
 
 import AppLayout from '../../../lib/components/app-layout';
-import AppLayoutToolbar from '../../../lib/components/app-layout-toolbar';
 
 import classicStyles from '../../../lib/components/app-layout/styles.selectors.js';
 import refreshStyles from '../../../lib/components/app-layout/visual-refresh/styles.selectors.js';
@@ -51,15 +50,6 @@ test('AppLayout should not render content during SSR', () => {
     <AppLayout content="SSR content region" breadcrumbs="SSR breadcrumbs" navigation="SSR navigation" />
   );
   expect(content).not.toContain('SSR content region');
-  expect(content).toContain('SSR breadcrumbs');
-  expect(content).toContain('SSR navigation');
-});
-test('AppLayoutToolbar should render content, breadcrumbs, and navigation during SSR', () => {
-  globalWithFlags[Symbol.for('awsui-visual-refresh-flag')] = () => true;
-  const content = renderToStaticMarkup(
-    <AppLayoutToolbar content="SSR content region" breadcrumbs="SSR breadcrumbs" navigation="SSR navigation" />
-  );
-  expect(content).toContain('SSR content region');
   expect(content).toContain('SSR breadcrumbs');
   expect(content).toContain('SSR navigation');
 });
